requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
decorator==5.1.1
exceptiongroup==1.1.1
fabric==3.1.0
fake==0.8
Faker==18.10.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
invoke==2.1.2
outcome==1.2.0
packaging==23.1
paramiko==3.2.0
pluggy==1.0.0
pycparser==2.21
PyNaCl==1.5.0
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0
